[SIX-178] refactor : feign 모듈 생성 및 패키지 의존성 수정
