+++ onedayhero-api/build.gradle
@@ -12,9 +12,7 @@ jar {
 
 // 전역 변수 설정
 ext {
-    jjwtVersion = '0.11.5'
     snippetsDir = file('build/generated-snippets')
-    set('springCloudVersion', "2022.0.3")
 }
 
 
@@ -29,11 +27,6 @@ dependencies {
     // RESTDOCS
     asciidoctorExt 'org.springframework.restdocs:spring-restdocs-asciidoctor'
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
-
-    //JWT
-    runtimeOnly "io.jsonwebtoken:jjwt-jackson:${jjwtVersion}"
-    runtimeOnly "io.jsonwebtoken:jjwt-impl:${jjwtVersion}"
-    implementation "io.jsonwebtoken:jjwt-api:${jjwtVersion}"
 }
 
 
@@ -71,11 +64,4 @@ task copyAsciidoc(type: Copy) {
 
 build {
     dependsOn copyAsciidoc
-}
-
-// OPEN FEIGN
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
 }
