--- onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/auth/response/LoginResponse.java
@@ -1,10 +1,0 @@
-package com.sixheroes.onedayheroapi.auth.response;
-
-import lombok.Builder;
-
-@Builder
-public record LoginResponse(
-        Long userId,
-        String accessToken
-) {
-}
