+++ onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/auth/response/oauth/KakaoAuthorizationCodeResponse.java
@@ -1,0 +1,7 @@
+package com.sixheroes.onedayheroapi.auth.response.oauth;
+
+public record KakaoAuthorizationCodeResponse(
+        String code,
+        String error
+) {
+}
