+++ onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/global/argumentsresolver/authuser/AuthUser.java
@@ -1,0 +1,12 @@
+package com.sixheroes.onedayheroapi.global.argumentsresolver.authuser;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthUser {
+
+}
