--- onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/global/auth/AuthUser.java
@@ -1,12 +1,0 @@
-package com.sixheroes.onedayheroapi.global.auth;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthUser {
-
-}
