--- onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/global/jwt/JwtProperties.java
@@ -1,17 +1,0 @@
-package com.sixheroes.onedayheroapi.global.jwt;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Getter
-@RequiredArgsConstructor
-@ConfigurationProperties(prefix = "jwt")
-public class JwtProperties {
-
-    private final String secretKey;
-    private final Long accessTokenExpiryTimeMs;
-    private final Long refreshTokenExpiryTimeMs;
-    private final String claimId;
-    private final String claimRole;
-}
