+++ onedayhero-application/build.gradle
@@ -1,22 +1,19 @@
 ext {
-    set('springCloudVersion', "2022.0.3")
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
+    jjwtVersion = '0.11.5'
 }
 
 dependencies {
+    //JWT
+    runtimeOnly "io.jsonwebtoken:jjwt-jackson:${jjwtVersion}"
+    runtimeOnly "io.jsonwebtoken:jjwt-impl:${jjwtVersion}"
+    api "io.jsonwebtoken:jjwt-api:${jjwtVersion}"
+
     //AWS S3
     implementation 'com.amazonaws:aws-java-sdk-s3:1.12.575'
     implementation project(':onedayhero-common')
     implementation project(':onedayhero-domain')
     implementation project(':onedayhero-infra-querydsl')
-
-    // OPEN FEIGN
-    implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
+    implementation project(':onedayhero-infra-openfeign')
 }
 
 tasks.named('test') {
@@ -30,4 +27,3 @@ jar {
 bootJar {
     enabled = false
 }
-
