+++ onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/global/configuration/OauthConfiguration.java
@@ -1,10 +1,10 @@
 package com.sixheroes.onedayheroapplication.global.configuration;
 
-import com.sixheroes.onedayheroapplication.global.oauth.kakao.OauthKakaoProperties;
+import com.sixheroes.onedayheroapplication.oauth.OauthProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
-@EnableConfigurationProperties(OauthKakaoProperties.class)
+@EnableConfigurationProperties(OauthProperties.class)
 @Configuration
 public class OauthConfiguration {
 }
