--- onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/global/oauth/OauthClientService.java
@@ -1,11 +1,0 @@
-package com.sixheroes.onedayheroapplication.global.oauth;
-
-
-public interface OauthClientService {
-
-    String getSocialType();
-
-    String requestToken(String code);
-
-    String requestResource(String accessToken);
-}
