--- onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/global/oauth/kakao/response/KakaoAuthorizationCodeResponse.java
@@ -1,7 +1,0 @@
-package com.sixheroes.onedayheroapplication.global.oauth.kakao.response;
-
-public record KakaoAuthorizationCodeResponse(
-        String code,
-        String error
-) {
-}
