--- onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/global/oauth/kakao/response/KakaoTokenResponse.java
@@ -1,11 +1,0 @@
-package com.sixheroes.onedayheroapplication.global.oauth.kakao.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record KakaoTokenResponse(
-        @JsonProperty("access_token")
-        String accessToken
-) {
-}
