+++ onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/oauth/OauthClientService.java
@@ -1,0 +1,11 @@
+package com.sixheroes.onedayheroapplication.oauth;
+
+
+public interface OauthClientService {
+
+    String getSocialType();
+
+    String requestToken(String code);
+
+    String requestResource(String accessToken);
+}
