+++ onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/oauth/OauthLoginService.java
@@ -1,0 +1,10 @@
+package com.sixheroes.onedayheroapplication.oauth;
+
+import com.sixheroes.onedayheroapplication.user.response.UserAuthResponse;
+
+public interface OauthLoginService {
+
+    Boolean supports(String authorizationServer);
+
+    UserAuthResponse login(String code);
+}
