+++ onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/oauth/response/LoginResponse.java
@@ -1,0 +1,8 @@
+package com.sixheroes.onedayheroapplication.oauth.response;
+
+
+public record LoginResponse(
+        Long userId,
+        String accessToken
+) {
+}
