+++ onedayhero-infra-openfeign/src/main/java/com/sixheroes/onedayheroinfraopenfeign/configuration/FeignConfiguration.java
@@ -1,0 +1,9 @@
+package com.sixheroes.onedayheroinfraopenfeign.configuration;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@EnableFeignClients(basePackages = "com.sixheroes.onedayheroinfraopenfeign.feign")
+@Configuration
+public class FeignConfiguration {
+}
