+++ onedayhero-infra-openfeign/src/main/java/com/sixheroes/onedayheroinfraopenfeign/feign/client/oauth/kakao/KakaoAuthServerFeignClient.java
@@ -1,0 +1,21 @@
+package com.sixheroes.onedayheroinfraopenfeign.feign.client.oauth.kakao;
+
+import com.sixheroes.onedayheroinfraopenfeign.feign.client.oauth.kakao.response.KakaoTokenResponse;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import java.util.Map;
+
+
+@FeignClient(
+        name = "kakao-auth",
+        url = "https://kauth.kakao.com/oauth/token"
+)
+public interface KakaoAuthServerFeignClient {
+
+    @PostMapping(consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+    KakaoTokenResponse requestToken(Map<String, ?> data);
+}
+
+
