+++ onedayhero-infra-openfeign/src/main/java/com/sixheroes/onedayheroinfraopenfeign/feign/client/oauth/kakao/response/KakaoTokenResponse.java
@@ -1,0 +1,11 @@
+package com.sixheroes.onedayheroinfraopenfeign.feign.client.oauth.kakao.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record KakaoTokenResponse(
+        @JsonProperty("access_token")
+        String accessToken
+) {
+}
