+++ onedayhero-infra-openfeign/src/test/java/com/sixheroes/onedayheroinfraopenfeign/OnedayheroInfraOpenfeignApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sixheroes.onedayheroinfraopenfeign;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+class OnedayheroInfraOpenfeignApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
