+++ settings.gradle
@@ -4,3 +4,4 @@ include 'onedayhero-application'
 include 'onedayhero-common'
 include 'onedayhero-domain'
 include 'onedayhero-infra-querydsl'
+include 'onedayhero-infra-openfeign'
